+++ system/src/Grav/Framework/Route/Route.php
@@ -252,6 +252,25 @@ class Route
         return $this->withParam('queryParams', $param, $value);
     }
 
+    public function withoutParams()
+    {
+        return $this->withoutGravParams()->withoutQueryParams();
+    }
+
+    public function withoutGravParams()
+    {
+        $this->gravParams = [];
+
+        return $this;
+    }
+
+    public function withoutQueryParams()
+    {
+        $this->queryParams = [];
+
+        return $this;
+    }
+
     /**
      * @return \Grav\Framework\Uri\Uri
      */
@@ -300,7 +319,7 @@ class Route
             return $this;
         }
 
-        $new = clone $this;
+        $new = $this->copy();
         if ($value === null) {
             unset($new->{$type}[$param]);
         } else {
@@ -308,6 +327,11 @@ class Route
         }
 
         return $new;
+    }
+
+    protected function copy()
+    {
+        return clone $this;
     }
 
     /**